+++ src/main/java/com/drmangotea/tfmg/base/fluid/ForceableFluidTank.java
@@ -1,16 +1,19 @@
 package com.drmangotea.tfmg.base.fluid;
 
 import com.simibubi.create.foundation.fluid.SmartFluidTank;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
 import net.neoforged.neoforge.fluids.FluidStack;
 
+import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
 
 /**
  * Simple tank with insertion and extraction restrictions that can be bypassed manually.
  */
+@MethodsReturnNonnullByDefault @ParametersAreNonnullByDefault
 public class ForceableFluidTank extends SmartFluidTank {
     private boolean extractionAllowed = true, insertionAllowed = true;